+++ src/RequestBuilder/EventsRequestBuilder.php
@@ -4,6 +4,7 @@ namespace Lmc\Matej\RequestBuilder;
 
 use Fig\Http\Message\RequestMethodInterface;
 use Lmc\Matej\Exception\LogicException;
+use Lmc\Matej\Model\Assertion;
 use Lmc\Matej\Model\Command\AbstractCommand;
 use Lmc\Matej\Model\Command\Interaction;
 use Lmc\Matej\Model\Command\ItemProperty;
@@ -82,6 +83,7 @@ class EventsRequestBuilder extends AbstractRequestBuilder
         if (empty($this->commands)) {
             throw new LogicException('At least one command must be added to the builder before sending the request');
         }
+        Assertion::batchSize($this->commands);
 
         return new Request(self::ENDPOINT_PATH, RequestMethodInterface::METHOD_POST, $this->commands, $this->requestId);
     }